Cardano.Benchmarking.Topology

Documentation

data Topology

A topology as it's used to define benchmarking profiles.

Constructors

Topology

Fields

coreNodes :: [Node]

relayNodes :: [Node]

data Node

A node as it's used to define benchmarking profiles. These nodes don't have remote addresses allocated, are referenced by name.

Constructors

Node

Fields

name :: String

nodeId :: Int

Placement, not present in Cardano.Node.Configuration.Topology.

region :: Location

As Cardano.Node.Configuration.Topology but names not addresses.

producers :: [String]

Actually always IOHK.

org :: String

pools :: Maybe Int

stakePool :: Maybe Bool

data Location

Location is either "loopback" used for local runs or an AWS Region name used for cloud runs.

Constructors

Loopback

AWS AWSRegion

data AWSRegion

The AWS Regions we support (No Availability Zone, the "c" in "us-east-2c").

Constructors

AP_SOUTHEAST_2

EU_CENTRAL_1

US_EAST_1

US_EAST_2

data CoreNodesParams

Four types of topologies for the core nodes.

Constructors

Line

Fields

tpSize :: Int

Number of nodes to use.

tpLocation :: Location

Unique location to use.

tpIdPools :: Int -> Maybe Int

Node i pool number.

UniCircle

Fields

tpSize :: Int

Number of nodes to use.

tpLocation :: Location

Unique location to use.

tpIdPools :: Int -> Maybe Int

Node i pool number.

Torus

Fields

tpSize :: Int

Number of nodes to use.

tpLocations :: [Location]

Locations to use.

tpIdPools :: Int -> Maybe Int

Node i pool number.

TorusDense

Fields

tpSize :: Int

Number of nodes to use.

tpLocations :: [Location]

Locations to use.

tpIdPools :: Int -> Maybe Int

Node i pool number.

mkTopology :: CoreNodesParams -> Maybe Location -> Topology

Create a Topology, optional explorer node that connects to all core nodes.

mkCoreNodes :: CoreNodesParams -> [Node]

mkExplorer :: Location -> [Node] -> Node